+++ misc_utils/__init__.py
@@ -8,6 +8,16 @@ from .misc_utils import color_print
 from .misc_utils import print_args
 
 from .misc_utils import get_logger
+from .misc_utils import cmd
+
+
+######################
+#        math
+######################
+from .misc_utils import hash
+from .misc_utils import gambling
+from .misc_utils import mean
+
 
 ######################
 #      safe load
@@ -28,6 +38,12 @@ from .misc_utils import try_make_dir
 from .misc_utils import get_file_name
 from .misc_utils import get_dir_name
 from .misc_utils import get_file_paths_by_pattern
+
+from .misc_utils import file_lines
+from .misc_utils import save_pickle
+from .misc_utils import load_pickle
+from .misc_utils import save_json
+from .misc_utils import load_json
 
 
 #######################
@@ -62,4 +78,5 @@ from .misc_utils import is_file_image
 #######################
 #      decorators
 #######################
-from .decorators import get_timer
+from .decorators import timer
+from .decorators import deprecated